% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service_db.R
\name{db_sync_project}
\alias{db_sync_project}
\title{Sync local project to database}
\usage{
db_sync_project(project_id, con = NULL)
}
\arguments{
\item{project_id}{Character. Local project ID.}

\item{con}{DBI connection. If NULL, creates a new one.}
}
\value{
Logical. TRUE if successful.
}
\description{
Uploads a local project (parquet files) to the PostGIS database.
}
