% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service_compute.R
\name{download_functions}
\alias{download_functions}
\title{Download functions for external data sources}
\description{
Functions to download raster and vector data from various French
geographic data providers (IGN, INPN, Copernicus).
}
\keyword{internal}
