% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_progress.R
\name{mod_progress}
\alias{mod_progress}
\title{Progress Module for nemetonApp}
\description{
Shiny module for displaying computation progress.
Shows download and computation status with detailed progress.
}
\keyword{internal}
