% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_i18n.R
\name{utils_i18n}
\alias{utils_i18n}
\title{Internationalization (i18n) for nemetonApp}
\description{
Translation system for the nemetonApp Shiny application.
Supports French (fr) and English (en).
}
\keyword{internal}
